Support instances of services
Create an instance from an S6_RC_TEMPLATE. The name of the dynamic service is "instance@template"
The template from the directory /etc/s6-rc/templates/ is copied with the whole service name
to "/run/service". An envfile called "env/instance" is created inside the service with 2 variables
 - INSTANCE=<instance-name>
 - TEMPLATE=<template-name>

The "run" and "finish" scripts may source it or the execlie way: *envfile env/instance* it and
use the INSTANCE variable name for starting this specific instance.
An optional iexecutablefile in the template: *data/check-instance* may be used to validate
the instance name. It will be called with the instance-name as argument and must exit 0
if the instance is valid. If it exits non-zero, the instance creation fails.