Automatically connect consumer-for and producer-for
The s6-rc-compile program requires matching producer-for
and consumer-for entries.

This is not ideal for unidirectional dependencies.
Whenever a producer and consumer are in different layers,
they both depend on each other.

With this change it is sufficient to name the peer either in the
S6RC_LONGRUN_xxxx[consumer-for] or S6RC_LONGRUN_xxxx[producer-for].
The postinst script will create the corresponding entry in the peer
if it is missing