Add fdstore for sockets and template pipes
Do not use the s6rc-fdholder service, which is not intended to
be used by other fd than those managed by s6rc.

The lighttpd_sockets did depend on the s6rc-fdholder, but a dependency
could not be expressed (an other hint that a separate
fd-holder service should be used).

An fd-holder service called "fdstore" now collects sockets
for socketactivation and pipes for "funnel loggers" of templates.
The socket/pipe creation and storage is performed in a "data/socket"
execline script of each service that uses it.

An fdstore-fill oneshot executes the socket scripts early in the
boot process. Both services are part of the "fdstorage" bundle.

Templates with an "all-in-one" logger also use this technology.
The fdstorage can be used for socketactivation mechanisms.