Fix sysctl oneshot not allowing more than one file
The "sysctl" oneshot loads additional sysctl settings from
/etc/sysctl.d/*.conf. Before running the actual command to load these
settings, it tests if there are any files that match the pattern.
The globbing is done with "elglob" which splits the result into as many
words as there are matches for the pattern. This is the required
behavior when using "sysctl -p". During the test, however, an error
occurs if there is more than one match, as "test" only accepts one
argument.

Use forx to handle 0,1 and more sysctl calls.
Inspired by PR#4 by Lennart Koschick <lk2601@icloud.com>